logic: compilationdatabase parsing
* compilationdatabase cleaned
* add SolutionparserCompliationDatabase
* added cmake_build_type_init as release for better autocomplition on linux
* external source handling for different linux versions
* clean script remove windows symbolic links only on windows
* linux external lib handling for different build easier

+++ script/clean.sh
@@ -2,15 +2,28 @@
 
 SUCCESS="\033[32mSuccess:\033[00m"
 
+# Determine current platform
+PLATFORM='unknown'
+if [ "$(uname)" == "Darwin" ]; then
+	PLATFORM='MacOS'
+elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
+	PLATFORM='Linux'
+elif [ "$(expr substr $(uname -s) 1 10)" == "MINGW32_NT" ]; then
+	PLATFORM='Windows'
+fi
+
 # Enter masterproject directory
 MY_PATH=`dirname "$0"`
 cd $MY_PATH/..
 
-# Remove symbolic links
-cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Debug\data' &
-cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Debug\user' &
-cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Release\data' &
-cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Release\user' &
+if [ $PLATFORM == "Windows" ]; then
+    # Remove symbolic links
+    cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Debug\data' &
+    cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Debug\user' &
+    cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Release\data' &
+    cmd //c 'rmdir '.$MY_PATH.'\..\bin\app\Release\user' &
+
+fi
 
 # Remove folders and contents
 rm -rf build
